// shared/api/generated/model/propertyInfo.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Dingle Swagger API
 * Dingle API
 * OpenAPI spec version: v1.0.0
 */
import type { PropertyInfoOrientation } from './propertyInfoOrientation';
import type { PropertyInfoPropertyType } from './propertyInfoPropertyType';

export interface PropertyInfo {
  address?: string;
  apartmentName?: string;
  bathrooms?: number;
  bedrooms?: number;
  evParkingSpaces?: number;
  exclusiveArea?: number;
  floor?: number;
  latitude?: number;
  liked?: boolean;
  longitude?: number;
  orientation?: PropertyInfoOrientation;
  parkingRatio?: number;
  propertyType?: PropertyInfoPropertyType;
  supplyArea?: number;
  totalFloor?: number;
}